Added a simple event logging function to Block.
This commit adds the log() function in Block.pm, allowing subclasses of Block to easily log events in the database. Note that this is
distinct from Logging.pm - the Logging.pm module is intended to be used for internal (generally debugging or error/fatal) logging to
STDERR or files, not the day-to-day recording of events that might be needed for auditing.