Added support for language changing.
This should actually address FS#70 - by loading a default language on creation, and then allowing another language to be loaded later and
overwrite some or all of the previously loaded language, this allows for fallbacks to defaults, or complete clearing if needed.

Note that langdir can't be changed, that can only be set at creation.